+++ src/template.html
@@ -29,7 +29,7 @@
       </div>
       <div class="task-list">
         <form class="form-input">
-          <input type="text" placeholder="Today's to do" />
+          <input type="text" id="todo" placeholder="Today's to do" />
           <button class="send"><ion-icon name="send"></ion-icon></button>
         </form>
       </div>
