+++ src/todos.js
@@ -1,0 +1,2 @@
+const todos = JSON.parse(localStorage.getItem('todo-list')) || [];
+export default todos;
